simplified client file, renamed files
